Refactor cache classes and add get method
The commit refactors the FIFOCache and LIFOCache classes to include a get method that retrieves a cache value. The get method takes a key as an argument and returns the corresponding value from the cache. The commit also includes a test file that tests the functionality of the get method.