Refactor LRU cache implementation
This commit refactors the implementation of the Least Recently Used (LRU) cache in the `LRUCache` class. The `decrement` method has been updated to handle the case where the key being decremented is the most recently used key. Additionally, some minor changes have been made to the `put` method to improve readability. The `__main__` block has been removed as it was only used for testing purposes.