Refactor LRU Cache implementation
This commit refactors the implementation of the Least Recently Used (LRU) Cache by updating the `decrement` method to use a more descriptive variable name and fixing a bug where the `mini` variable was not being updated correctly. Additionally, the `put` and `get` methods were updated to use the new variable name. The `3-main.py` file was also updated to test the LRU Cache implementation with a larger number of items and to verify that the cache correctly discards the least recently used items when the cache is full.