Refactor LRUCache class to use access_order instead of points
This commit refactors the LRUCache class to use access_order instead of points to keep track of the order in which keys were accessed in the cache. This change improves the readability and maintainability of the code. The decrement method has also been updated to work with the new access_order dictionary.